Organize the file structure for yellow-ribbon
