--- src/applications/yellow-ribbon/reducers/yellowRibbonReducer.js
@@ -1,40 +1,0 @@
-// Relative imports.
-import {
-  FETCH_RESULTS,
-  FETCH_RESULTS_FAILURE,
-  FETCH_RESULTS_SUCCESS,
-  UPDATE_PAGINATION,
-  UPDATE_RESULTS,
-} from '../constants';
-
-const initialState = {
-  error: '',
-  fetching: false,
-  page: 1,
-  query: '',
-  results: null,
-  startIndex: 0,
-};
-
-export default (state = initialState, action) => {
-  switch (action.type) {
-    case FETCH_RESULTS: {
-      return { ...state, error: '', fetching: true, query: action.query };
-    }
-    case FETCH_RESULTS_FAILURE: {
-      return { ...state, error: action.error, fetching: false };
-    }
-    case FETCH_RESULTS_SUCCESS: {
-      return { ...state, fetching: false, results: action.results };
-    }
-    case UPDATE_PAGINATION: {
-      return { ...state, page: action.page, startIndex: action.startIndex };
-    }
-    case UPDATE_RESULTS: {
-      return { ...state, results: action.results };
-    }
-    default: {
-      return { ...state };
-    }
-  }
-};
