--- src/applications/yellow-ribbon/tests/components/Card.unit.spec.jsx
@@ -1,25 +1,0 @@
-// Dependencies.
-import React from 'react';
-import { expect } from 'chai';
-import { shallow } from 'enzyme';
-// Relative imports.
-import Card from '../../components/Card';
-
-describe('Find VA Results <Card>', () => {
-  it('should render', () => {
-    const props = {
-      heading: 'heading',
-      href: '/href/',
-      description: 'description',
-    };
-
-    const tree = shallow(<Card {...props} />);
-    const text = tree.text();
-
-    expect(tree.find('a').prop('href')).to.equal(props.href);
-    expect(text).to.include(props.heading);
-    expect(text).to.include(props.description);
-
-    tree.unmount();
-  });
-});
