--- src/applications/yellow-ribbon/tests/components/TopTasks.unit.spec.jsx
@@ -1,16 +1,0 @@
-// Dependencies.
-import React from 'react';
-import { expect } from 'chai';
-import { shallow } from 'enzyme';
-// Relative imports.
-import TopTasks from '../../components/TopTasks';
-
-describe('Find VA Results <TopTasks>', () => {
-  it('should render', () => {
-    const tree = shallow(<TopTasks />);
-
-    expect(tree.find('h2').text()).to.contain('Top tasks');
-
-    tree.unmount();
-  });
-});
