+++ src/applications/yellow-ribbon/yellow-ribbon-entry.jsx
@@ -1,6 +1,7 @@
+// Styling imports.
 import 'platform/polyfills';
 import './sass/yellow-ribbon.scss';
-
+// Relative imports.
 import startApp from 'platform/startup';
 import routes from './routes';
 import manifest from './manifest.json';
